Add a new type for MissingTIROperationTermination for reconciliation
In the case when the termination info is missing, we can not send
the full data to the customs for them to fill it (we only have
the carnet number, the page and the country). So we need to introduce
a new type to hold just this info.
Note that it makes the reason redundant, but it is safer to leave
it in.